Add OSGi plugin for use by git client plugin
The git client plugin bundles JGit so that Jenkins consumers of the JGit
API can use the JGit version that is already included in the git client
plugin.  As part of https://issues.jenkins.io/browse/JENKINS-69159 and
https://github.com/jenkinsci/git-client-plugin/pull/956 the git client
plugin is switching to use Apache Mina ssh.

The OSGi jar file is included in the git client plugin as a transitive
dependency with https://github.com/jenkinsci/git-client-plugin/pull/956 .
I think it will be better to make it an explicit API plugin rather than
having it included as a transitive dependency in the git client plugin.