remove trilead dep, add optional dep to BC and eddsa
Plugin depending on trilead (a different ssh implementaion)
makes no logical sense.

Add optional dependencies to bouncycasle and eddsa plugin.

The latter is needed for EdDSA keys, the former can add extra things
from BouncyCastle.

This prevents some classloading excpeptions where the context
classloader is set to a class that can see bouncycastle or eddsa code
yet the classloader of code in these plugin can not.

Now if eddsa or bouncycastle plugins are installed this plugin will
consistently be able to see the code.

remove management of ssh-credentials plugin